[refactor]:Main #73 코드 정리
