+++ MainService/src/main/java/com/kusitms/mainservice/domain/mypage/dto/MySharedContentDto.java
@@ -1,6 +1,6 @@
-package com.kusitms.mainservice.global.infra.mypage.dto;
+package com.kusitms.mainservice.domain.mypage.dto;
 
-import com.kusitms.mainservice.global.infra.mypage.domain.SharedType;
+import com.kusitms.mainservice.domain.mypage.domain.SharedType;
 import lombok.Builder;
 import lombok.Getter;
