+++ MainService/src/main/java/com/kusitms/mainservice/domain/mypage/service/MyPageService.java
@@ -1,5 +1,8 @@
-package com.kusitms.mainservice.global.infra.mypage.service;
+package com.kusitms.mainservice.domain.mypage.service;
 
+import com.kusitms.mainservice.domain.mypage.dto.MyPageResponseDto;
+import com.kusitms.mainservice.domain.mypage.dto.MyPageUserResponseDto;
+import com.kusitms.mainservice.domain.mypage.dto.MySharedContentDto;
 import com.kusitms.mainservice.domain.roadmap.domain.Roadmap;
 import com.kusitms.mainservice.domain.roadmap.repository.RoadmapDownloadRepository;
 import com.kusitms.mainservice.domain.roadmap.repository.RoadmapRepository;
@@ -8,11 +11,9 @@ import com.kusitms.mainservice.domain.template.repository.TemplateDownloadReposi
 import com.kusitms.mainservice.domain.template.repository.TemplateRepository;
 import com.kusitms.mainservice.domain.user.domain.User;
 import com.kusitms.mainservice.domain.user.dto.response.DetailUserResponseDto;
-import com.kusitms.mainservice.domain.user.service.UserService;
-import com.kusitms.mainservice.global.infra.mypage.domain.SharedType;
-import com.kusitms.mainservice.global.infra.mypage.dto.MyPageResponseDto;
-import com.kusitms.mainservice.global.infra.mypage.dto.MyPageUserResponseDto;
-import com.kusitms.mainservice.global.infra.mypage.dto.MySharedContentDto;
+import com.kusitms.mainservice.domain.user.repository.UserRepository;
+import com.kusitms.mainservice.global.error.exception.EntityNotFoundException;
+import com.kusitms.mainservice.domain.mypage.domain.SharedType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageImpl;
@@ -25,7 +26,7 @@ import org.springframework.data.domain.Pageable;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.kusitms.mainservice.global.infra.mypage.domain.SharedType.getEnumSharedTypeFromStringSharedType;
+import static com.kusitms.mainservice.global.error.ErrorCode.USER_NOT_FOUND;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -36,8 +37,8 @@ public class MyPageService {
     private final RoadmapDownloadRepository roadmapDownloadRepository;
     private final TemplateRepository templateRepository;
     private final RoadmapRepository roadmapRepository;
-    private final UserService userService;
-    public MyPageResponseDto getMyPageResponse(Long userId,Pageable pageable ){
+    private final UserRepository userRepository;
+    public MyPageResponseDto getMyPageResponse(Long userId, Pageable pageable ){
         MyPageResponseDto myPageResponseDto = createMyPageResponseDto(userId, pageable);
     return myPageResponseDto;
     }
@@ -79,11 +80,28 @@ public class MyPageService {
         return resultPage;
     }
     private MyPageUserResponseDto createMyPageUserResponseDto(Long userId){
-        User user = userService.getUserByUserId(userId);
-        DetailUserResponseDto detailUserResponseDto = userService.createDetailUserResponseDto(user);
+        User user = getUserByUserId(userId);
+        DetailUserResponseDto detailUserResponseDto = createDetailUserResponseDto(user);
         int templateNum = detailUserResponseDto.getTemplateNum();
         int roadmapNum = detailUserResponseDto.getRoadmapNum();
         int point = 0;
         return MyPageUserResponseDto.of(user,templateNum,roadmapNum,point);
+    }
+
+    private User getUserByUserId(Long id){
+        return userRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(USER_NOT_FOUND));
+    }
+
+    public DetailUserResponseDto createDetailUserResponseDto(User user){
+        int templateNum = getTemplateCountByUser(user);
+        int roadmapNum = getRoadmapCountByUser(user);
+        return DetailUserResponseDto.of(user, templateNum,roadmapNum);
+    }
+    private int getTemplateCountByUser(User user) {
+        return templateRepository.countByUser(user);
+    }
+    private int getRoadmapCountByUser(User user) {
+        return roadmapRepository.countByUser(user);
     }
 }
