+++ MainService/src/main/java/com/kusitms/mainservice/domain/user/controller/UserController.java
@@ -5,7 +5,9 @@ import com.kusitms.mainservice.domain.user.dto.request.UserSignInRequestDto;
 import com.kusitms.mainservice.domain.user.dto.request.UserSignUpRequestDto;
 import com.kusitms.mainservice.domain.user.dto.response.UserAuthResponseDto;
 import com.kusitms.mainservice.domain.user.service.AuthService;
+import com.kusitms.mainservice.domain.user.service.UserService;
 import com.kusitms.mainservice.global.common.SuccessResponse;
+import com.kusitms.mainservice.global.config.auth.UserId;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
